DNN-10507: Merged ResultCapturingActionInvoker Tests
